+++ backend/integration_tests/src/client/local_user_index.rs
@@ -1,0 +1,7 @@
+use crate::generate_update_call;
+use local_user_index_canister::*;
+
+// Queries
+
+// Updates
+generate_update_call!(join_group);