+++ backend/integration_tests/src/lib.rs
@@ -8,6 +8,7 @@ mod client;
 mod delete_group_tests;
 mod delete_message_tests;
 mod freeze_group_tests;
+mod join_group_tests;
 mod last_online_date_tests;
 mod notification_tests;
 mod poll_tests;