+++ frontend/openchat-agent/src/services/openchatAgent.ts
@@ -13,6 +13,8 @@ import { getAllUsers } from "../utils/userCache";
 import { UserIndexClient } from "./userIndex/userIndex.client";
 import { UserClient } from "./user/user.client";
 import { GroupClient } from "./group/group.client";
+import type { ILocalUserIndexClient } from "./localUserIndex/localUserIndex.client.interface";
+import { LocalUserIndexClient } from "./localUserIndex/localUserIndex.client";
 import type { INotificationsClient } from "./notifications/notifications.client.interface";
 import { NotificationsClient } from "./notifications/notifications.client";
 import type { IOnlineClient } from "./online/online.client.interface";
@@ -132,7 +134,7 @@ import {
 } from "openchat-shared";
 import type { Principal } from "@dfinity/principal";
 import { applyOptionUpdate } from "../utils/mapping";
-import {waitAll} from "../utils/promise";
+import { waitAll } from "../utils/promise";
 
 export const apiKey = Symbol();
 
@@ -215,6 +217,10 @@ export class OpenChatAgent extends EventTarget {
             return this._userClient;
         }
         throw new Error("Attempted to use the user client before it has been initialised");
+    }
+
+    private createLocalUserIndexClient(canisterId: string): ILocalUserIndexClient {
+        return LocalUserIndexClient.create(this.identity, this.config, canisterId);
     }
 
     private getProvidedInviteCode(chatId: string): string | undefined {
@@ -1057,8 +1063,14 @@ export class OpenChatAgent extends EventTarget {
         return this.userClient.leaveGroup(chatId);
     }
 
-    joinGroup(chatId: string): Promise<JoinGroupResponse> {
+    async joinGroup(chatId: string): Promise<JoinGroupResponse> {
         const inviteCode = this.getProvidedInviteCode(chatId);
+        // This can be enabled once userIndex, localUserIndex, user and group canisters have been upgraded
+        const useNewJoinGroup = false;
+        if (useNewJoinGroup) {
+            const localUserIndex = await this.getGroupClient(chatId).localUserIndex();
+            return this.createLocalUserIndexClient(localUserIndex).joinGroup(chatId, inviteCode);
+        }
         return this.userClient.joinGroup(chatId, inviteCode);
     }