+++ frontend/CHANGELOG.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Integrated the new `join_group` flow ([#2955](https://github.com/open-ic/open-chat/pull/2955)) (currently disabled) 
 - Allow authorized users to see deleted messages ([#2914](https://github.com/open-ic/open-chat/pull/2914)) (requires user and group canisters to be released first)
-- 
+
 ### Changed
+
+- Make the right hand panel dismissible via Esc or clicking away ([#2958](https://github.com/open-ic/open-chat/pull/2958))
 
 - Attempt to fix twitter card rendering ([#2957](https://github.com/open-ic/open-chat/pull/2957))